refactor: 애프터노트 편집기 구조 개선 및 중복 코드 정리
- `SaveAfternoteUseCase`를 `SaveAfternoteOrchestrator`로 이름을 변경하고 관련 로직 이동
- `Song` 모델 및 관련 컴포저블의 패키지 위치를 `memorial.playlist`로 이동하여 구조 정리
- `AlbumCover` 등 공용 모델과 Mock 데이터를 `core.model` 패키지로 이전하여 모듈 간 의존성 최적화
- `AfternoteDebugMockNetworkInterceptor` 내 하드코딩된 JSON 문자열들을 `AfternoteMockFixtures`로 분리하여 관리
- `AddSongScreen`, `AddSongViewModel` 등 기존의 개별 UI/ViewModel 파일들을 삭제하고 공용 컴포넌트 활용 구조로 변경
- UI 상태 모델(`MemorialGuidelineEditorContentParams`, `ProcessingMethodListParams`)에 `@Stable`, `@Immutable` 어노테이션 추가로 성능 최적화
- 불필요한 더미 데이터 클래스 및 사용되지 않는 파일 제거
