+++ feature/afternote/data/build.gradle.kts
@@ -10,5 +10,6 @@ dependencies {
     implementation(libs.androidx.datastore.preferences)
     implementation(projects.core.common)
     implementation(projects.core.domain)
+    implementation(projects.core.model)
     implementation(projects.feature.afternote.domain)
 }
