--- feature/afternote/presentation/src/main/kotlin/com/afternote/feature/afternote/presentation/author/editor/playlist/AddSongEvent.kt
@@ -1,10 +1,0 @@
-package com.afternote.feature.afternote.presentation.author.editor.playlist
-
-/**
- * UI events for Add Song screen.
- */
-sealed interface AddSongEvent {
-    data class SearchQueryChange(
-        val query: String,
-    ) : AddSongEvent
-}
