--- feature/afternote/presentation/src/main/kotlin/com/afternote/feature/afternote/presentation/author/editor/playlist/AddSongUiState.kt
@@ -1,13 +1,0 @@
-package com.afternote.feature.afternote.presentation.author.editor.playlist
-
-import com.afternote.feature.afternote.presentation.shared.model.PlaylistSongDisplay
-
-/**
- * UI state for Add Song screen (search-driven list from API).
- */
-data class AddSongUiState(
-    val songs: List<PlaylistSongDisplay> = emptyList(),
-    val searchQuery: String = "",
-    val isLoading: Boolean = false,
-    val errorMessage: String? = null,
-)
