--- feature/afternote/presentation/src/main/kotlin/com/afternote/feature/afternote/presentation/author/editor/usecase/SaveAfternoteUseCase.kt
@@ -1,93 +1,0 @@
-package com.afternote.feature.afternote.presentation.author.editor.usecase
-
-import com.afternote.feature.afternote.domain.repository.AfternoteRepository
-import com.afternote.feature.afternote.domain.usecase.editor.ResolveMemorialMediaForSaveUseCase
-import com.afternote.feature.afternote.presentation.author.editor.SaveAfternoteMemorialMedia
-import com.afternote.feature.afternote.presentation.author.editor.mapper.AfternoteEditorMapper
-import com.afternote.feature.afternote.presentation.author.editor.mapper.CreateInput
-import com.afternote.feature.afternote.presentation.author.editor.mapper.MemorialMediaUrls
-import com.afternote.feature.afternote.presentation.author.editor.model.EditorCategory
-import com.afternote.feature.afternote.presentation.author.editor.model.RegisterAfternotePayload
-import com.afternote.feature.afternote.presentation.author.editor.state.MemorialPlaylistStateHolder
-import javax.inject.Inject
-
-/**
- * 검증 이후 미디어 업로드·생성/수정 API 호출까지 담당 (매퍼·리포지토리 조합).
- */
-class SaveAfternoteUseCase
-    @Inject
-    constructor(
-        private val afternoteRepository: AfternoteRepository,
-        private val resolveMemorialMediaForSave: ResolveMemorialMediaForSaveUseCase,
-    ) {
-        suspend operator fun invoke(
-            editingId: Long?,
-            categoryForApi: EditorCategory,
-            payload: RegisterAfternotePayload,
-            selectedReceiverIds: List<Long>,
-            playlistStateHolder: MemorialPlaylistStateHolder?,
-            memorialMedia: SaveAfternoteMemorialMedia,
-        ): Result<Long> {
-            val resolved =
-                resolveMemorialMediaForSave(
-                    funeralVideoUrl = memorialMedia.funeralVideoUrl,
-                    memorialPhotoUrl = memorialMedia.memorialPhotoUrl,
-                    pickedMemorialPhotoUri = memorialMedia.pickedMemorialPhotoUri,
-                    funeralThumbnailUrl = memorialMedia.funeralThumbnailUrl,
-                    isUpdate = editingId != null,
-                ).getOrElse { return Result.failure(it) }
-
-            return if (editingId != null) {
-                val updatePayload =
-                    AfternoteEditorMapper.buildUpdatePayload(
-                        category = categoryForApi,
-                        payload = payload,
-                        selectedReceiverIds = selectedReceiverIds,
-                        playlistStateHolder = playlistStateHolder,
-                        memorialMedia =
-                            MemorialMediaUrls(
-                                funeralVideoUrl = resolved.videoUrlForUpdate,
-                                funeralThumbnailUrl = resolved.funeralThumbnailUrlForUpdate,
-                                memorialPhotoUrl = resolved.resolvedMemorialPhotoUrl,
-                            ),
-                    )
-                afternoteRepository.update(id = editingId, payload = updatePayload)
-            } else {
-                performCreate(
-                    category = categoryForApi,
-                    payload = payload,
-                    selectedReceiverIds = selectedReceiverIds,
-                    playlistStateHolder = playlistStateHolder,
-                    funeralVideoUrl = resolved.resolvedVideoUrl,
-                    funeralThumbnailUrl = memorialMedia.funeralThumbnailUrl,
-                    memorialPhotoUrl = resolved.resolvedMemorialPhotoUrl,
-                )
-            }
-        }
-
-        private suspend fun performCreate(
-            category: EditorCategory,
-            payload: RegisterAfternotePayload,
-            selectedReceiverIds: List<Long>,
-            playlistStateHolder: MemorialPlaylistStateHolder?,
-            funeralVideoUrl: String?,
-            funeralThumbnailUrl: String?,
-            memorialPhotoUrl: String?,
-        ): Result<Long> {
-            val createInput =
-                AfternoteEditorMapper.buildCreateInput(
-                    category = category,
-                    payload = payload,
-                    selectedReceiverIds = selectedReceiverIds,
-                    playlistStateHolder = playlistStateHolder,
-                    funeralVideoUrl = funeralVideoUrl,
-                    funeralThumbnailUrl = funeralThumbnailUrl,
-                    memorialPhotoUrl = memorialPhotoUrl,
-                )
-            return when (createInput) {
-                is CreateInput.Social -> afternoteRepository.createSocial(createInput.payload)
-                is CreateInput.Gallery -> afternoteRepository.createGallery(createInput.payload)
-                is CreateInput.Playlist -> afternoteRepository.createPlaylist(createInput.payload)
-            }
-        }
-    }
