--- feature/afternote/presentation/src/main/kotlin/com/afternote/feature/afternote/presentation/shared/model/dummy/AlbumDummies.kt
@@ -1,17 +1,0 @@
-package com.afternote.feature.afternote.presentation.shared.model.dummy
-
-import com.afternote.feature.afternote.presentation.shared.detail.song.AlbumCover
-
-/**
- * Default album cover list for playlists (Previews and dummy state).
- * Replace with real data when loaded from API.
- */
-object AlbumDummies {
-    val list: List<AlbumCover> =
-        listOf(
-            AlbumCover("1"),
-            AlbumCover("2"),
-            AlbumCover("3"),
-            AlbumCover("4"),
-        )
-}
