--- feature/afternote/presentation/src/main/kotlin/com/afternote/feature/afternote/presentation/shared/model/dummy/preview/TimeLetterPreviewItem.kt
@@ -1,6 +1,0 @@
-package com.afternote.feature.afternote.presentation.shared.model.dummy.preview
-
-data class TimeLetterPreviewItem(
-    val title: String,
-    val subtitle: String = "",
-)
